requirements.txt
appdirs==1.4.4
asttokens==2.0.5
attrs==22.1.0
Automat==20.2.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
cssselect==1.1.0
cycler==0.11.0
decorator==5.1.1
dnspython==2.0.0
executing==0.9.1
fake-useragent==0.1.11
filelock==3.7.1
fonttools==4.34.4
future==0.18.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
ipwhois==1.2.0
ipython==8.4.0
itemadapter==0.6.0
itemloaders==1.0.4
jedi==0.18.1
jmespath==1.0.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.23.1
packaging==21.3
pandas==1.4.3
parse==1.19.0
parsel==1.6.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
Protego==0.2.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.0
pycparser==2.21
PyDispatcher==2.0.5
pyee==8.2.2
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-whois==0.8.0
pytz==2022.1
PyX==0.15
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-html==0.10.0
rich==12.5.1
rich-click==1.5.2
scapy==2.4.5
Scrapy==2.6.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tldextract==3.3.1
toml==0.10.2
tqdm==4.64.0
traitlets==5.3.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
urllib3==1.26.11
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.3
zipp==3.8.1
zope.interface==5.4.0
